requirements.txt
fastapi
uvicorn
numpy
opencv-python
tensorflow
pillow
scikit-learn
python-dotenv
google-generativeai
python-multipart

